src/requirements.txt
Flask-SocketIO==5.4.1
Flask==3.1.0
Jinja2==3.1.4
MarkupSafe==3.0.2
Werkzeug==3.1.3
bidict==0.23.1
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
google-api-core==2.23.0
google-api-python-client==2.153.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-auth==2.36.0
googleapis-common-protos==1.66.0
h11==0.14.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
oauthlib==3.2.2
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.2.0
python-engineio==4.10.1
python-socketio==5.11.4
requests-oauthlib==2.0.0
requests==2.32.3
rsa==4.9
simple-websocket==1.1.0
uritemplate==4.1.1
urllib3==2.2.3
wsproto==1.2.0
